replaced entry spliter code with single call to String.prototype.split()
The previous code would take in the whole file and parse it byte by byte
keeping track of the last byte it parsed and a string the represented
the current entry, if it encounterd a newline (`\n`) and the last parsed
byte wasn't a blackslash (i.e. the newline wasn't escaped) it would
constructe an entry object from the current entry string and yielded it

The new code removes (or well sortens) the first part of the code
where we parse the history file spliting on unescaped newlines into a
single call to String.prototype.split(). Where we use it to split on a
regex that searches for newlines (with a negative backward lookup for a
backslash)
`/(?<!\\)\n/`
        `\n` | split on a newline
 `(?<!\\)`   | only if the character before wasn't a backslash